# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API gateway for the Gardening Agent system. Sits between Verdant (React frontend) and Rhizome (Python LangGraph agent). Handles authentication, encrypted provider key storage, and request routing.",
        "title": "Cambium — Gardening Agent API",
        "contact": {
            "name": "Gardening Agent"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat (non-streaming)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID (from POST /api/v1/threads)",
                        "name": "thread_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Rhizome unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/resume": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Resume interaction (non-streaming)",
                "parameters": [
                    {
                        "description": "Thread ID and resolution value",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ResumeRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/resume/stream": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Resume interaction (SSE streaming)",
                "parameters": [
                    {
                        "description": "Thread ID and resolution value",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ResumeRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of typed events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/stream": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Chat (SSE streaming)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID (from POST /api/v1/threads)",
                        "name": "thread_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "User message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of typed events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Only return notifications after this timestamp",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Notification stream (SSE)",
                "responses": {
                    "200": {
                        "description": "SSE stream of typed events (alert, interaction_pending, job_started/job_step/job_complete/job_failed, heartbeat)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/threads/{id}/session-context": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Get thread session context",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SessionContextView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Update thread session context",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial session context override",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateSessionContextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SessionContextView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ChatRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "What should I do in my garden today?"
                }
            }
        },
        "api.ChatResponse": {
            "type": "object",
            "properties": {
                "interaction": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "response": {
                    "type": "string",
                    "example": "Based on your garden profile..."
                },
                "thread_id": {
                    "type": "string",
                    "example": "silver-fern-cascade"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid or expired token"
                }
            }
        },
...
        "api.ResumeRequestBody": {
            "type": "object",
            "properties": {
                "resolution": {
                    "type": "string",
                    "example": "confirm"
                },
                "thread_id": {
                    "type": "string",
                    "example": "silver-fern-cascade"
                }
            }
        },
        "api.SessionContextView": {
            "type": "object",
            "properties": {
                "available_minutes": {
                    "type": "integer",
                    "example": 30
                },
                "energy_level": {
                    "type": "string",
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ],
                    "example": "low"
                },
                "focus_label": {
                    "type": "string",
                    "example": "Tomato bed refresh"
                },
                "focus_project_id": {
                    "type": "string",
                    "example": "project-uuid"
                },
                "open_to_outdoor_work": {
                    "type": "boolean",
                    "example": true
                },
                "preferred_location_type": {
                    "type": "string",
                    "enum": [
                        "bed",
                        "container"
                    ],
                    "example": "container"
                },
                "source": {
                    "type": "string",
                    "enum": [
                        "unset",
                        "inferred",
                        "user"
                    ],
                    "example": "inferred"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-06-21T16:44:56Z"
                },
                "wants_quick_wins": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "api.UpdateSessionContextRequest": {
            "type": "object",
            "properties": {
                "available_minutes": {
                    "type": "integer",
                    "example": 15
                },
                "energy_level": {
                    "type": "string",
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ],
                    "example": "medium"
                },
                "focus_project_id": {
                    "type": "string",
                    "example": "project-uuid"
                },
                "open_to_outdoor_work": {
                    "type": "boolean",
                    "example": false
                },
                "preferred_location_type": {
                    "type": "string",
                    "enum": [
                        "bed",
                        "container"
                    ],
                    "example": "bed"
                },
                "wants_quick_wins": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token — format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}